chore:
`chown` is now only done if there are files in the `settings` folder
that are not readable. This is necessary to avoid errors in a K8S
environment that occur when the mounted volume is readonly. In this
environment, the `defaultMode` for the volume must then be `0644` so
that the files can be read.