fix: avoid readable but not writeable files
It has happened that old announcements.db files, created as root, are now readable and not writable. This previously caused that their owner is not changed, but they also can not be opened and written.

As a solution, the option to bypass the user change with suitable rights is dropped. The owner is always changed. In addition, read and write permissions are given to the owner for safety.